fix duplicate symbol error in trace_api::configuration_utils
